requirements.txt
#pip freeze > requirements.txt pip3 freeze > requirements.txt.
aiohttp==3.7.4.post0
appier==1.21.7
async-timeout==3.0.1
attrs==21.2.0
AutoTS==0.4.1
backtest==0.7.6
Backtesting==0.3.3
backtrader==1.9.76.123
binance-api==0.1.2
bokeh==2.4.3
Brotli==1.0.9
bta-lib==
certifi==2021.5.30
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.4
clx-sdk-xms==1.0.0
colorama==0.4.4
cycler==0.11.0
Cython==0.29.23
dash==2.1.0
dash-bootstrap-components==1.0.3
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dateparser==1.0.0
ez-setup==0.9
Flask==2.0.3
Flask-Compress==1.10.1
fonttools==4.33.3
greenlet==1.1.2
idna==2.8
iso8601==0.1.14
itsdangerous==2.1.0
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.4.2
lxml==4.9.0
MarkupSafe==2.1.0
matplotlib==3.5.2
multidict==6.0.4
multitasking==0.0.10
notify-run==0.0.13
numpy==1.24.4
packaging==21.3
pandas==1.5.3
patsy==0.5.2
Pillow==9.1.1
plotly==5.6.0
psycopg2==2.9.3
pyine==1.1.2
PyJWT==1.7.1
pyparsing==3.0.9
PyQRCode==1.2.1
python-binance==1.0.12
python-dateutil==2.8.2
python-kucoin==2.1.2
pytz==2021.3
PyYAML==6.0
regex==2022.3.2
requests==2.28.0
retrying==1.3.3
scikit-learn==1.1.1
scipy==1.8.1
selenium==3.141.0
sinchsms==1.0.4
six==1.16.0
SQLAlchemy==1.4.32
statsmodels==0.13.2
tenacity==8.0.1
threadpoolctl==3.1.0
tornado==6.1
twilio==6.59.1
typing-extensions==3.10.0.0
tzlocal==2.1
ujson==5.8.0
urllib3==1.25.11
websockets==9.1
Werkzeug==2.0.3
yarl==1.9.2
yfinance==0.1.72
